+++ autopsy/reconfigure.py
@@ -387,6 +387,10 @@ class Parameter(object):
 class ConstrainedP(Parameter):
     """Parameter that is constrained by min and max values."""
 
+    # Constrain the value even further by looking at a linked variable.
+    _link = None
+
+
     def __init__(self, name, default, type, min = -2147483647, max = 2147483647, **kwargs):
         """Initialize a constrained parameter object.
 
@@ -450,6 +454,9 @@ class ConstrainedP(Parameter):
 
         if not self.min <= new_value <= self.max:
             raise ValueError("Value '%s' is not in %s <= value <= %s range." % (formatNumber(new_value), formatNumber(self.min), formatNumber(self.max)))
+
+        if self._link is not None:
+            new_value = self._link(new_value)
 
         Parameter.value.fset(self, new_value)
 
@@ -732,7 +739,9 @@ class ParameterReconfigure(object):
     def _reupdate(self):
         """Creates an update description of the parameters."""
 
-        _config = self._get_config(condition = lambda x: x.value != x.default)
+        # Note: Condition was removed because of the linked variables, as when
+        # the value should be moved to its default state, nothing would happen.
+        _config = self._get_config()#condition = lambda x: x.value != x.default)
         _config.groups = [
             GroupState(
                 name = "Default",
@@ -889,6 +898,22 @@ class ParameterServer(ParameterReconfigure):
         name -- name of the parameter, str
         """
         return name in self._parameters
+
+
+    def link(self, param1, param2):
+        """Links two constrained parameters together so one cannot be more then the other.
+
+        Arguments:
+        name1 -- name of the first ConstrainedP parameter, str
+        name2 -- name of the second ConstrainedP parameter, str
+        """
+
+        for param in (param1, param2):
+            if not isinstance(param, ConstrainedP):
+                raise TypeError("Parameter '%s' is of a type '%s'." % (param.name, type(param)))
+
+        param1._link = lambda x: min(x, param2.value)
+        param2._link = lambda x: max(x, param1.value)
 
 
     def update(self, parameters, only_existing = False):